+++ eras/conway/impl/src/Cardano/Ledger/Conway/Transition.hs
@@ -6,6 +6,7 @@
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE UndecidableSuperClasses #-}
@@ -33,7 +34,9 @@ import Cardano.Ledger.Conway.Translation ()
 import Cardano.Ledger.Conway.TxCert (Delegatee (..))
 import Cardano.Ledger.Core
 import Cardano.Ledger.Credential (Credential (..))
-import Cardano.Ledger.Shelley.Genesis (ShelleyGenesisStaking (..))
+import Cardano.Ledger.Shelley.Genesis (
+  ShelleyGenesisStaking (..),
+ )
 import Cardano.Ledger.Shelley.LedgerState (
   NewEpochState,
   curPParamsEpochStateL,
@@ -42,6 +45,7 @@ import Cardano.Ledger.Shelley.LedgerState (
   nesEsL,
  )
 import Cardano.Ledger.Shelley.Transition
+import Control.Monad.IO.Class (MonadIO)
 import Data.ListMap (ListMap)
 import qualified Data.ListMap as ListMap
 import qualified Data.Map.Strict as Map
@@ -76,9 +80,9 @@ instance EraTransition ConwayEra where
 
   mkTransitionConfig = ConwayTransitionConfig
 
-  injectIntoTestState cfg =
-    conwayRegisterInitialFundsThenStaking cfg
-      . alonzoInjectCostModels (cfg ^. tcPreviousEraConfigL . tcPreviousEraConfigL)
+  injectIntoTestState withFileHandle cfg newEpochState =
+    conwayRegisterInitialFundsThenStaking withFileHandle cfg $
+      alonzoInjectCostModels (cfg ^. tcPreviousEraConfigL . tcPreviousEraConfigL) newEpochState
 
   tcPreviousEraConfigL =
     lens ctcBabbageTransitionConfig (\ctc pc -> ctc {ctcBabbageTransitionConfig = pc})
@@ -104,20 +108,22 @@ tcInitialDRepsL =
 instance NoThunks (TransitionConfig ConwayEra)
 
 conwayRegisterInitialFundsThenStaking ::
-  ConwayEraTransition era =>
+  (ConwayEraTransition era, MonadIO m, MonadFail m) =>
+  WithFileHandle m ->
   TransitionConfig era ->
   NewEpochState era ->
-  NewEpochState era
-conwayRegisterInitialFundsThenStaking cfg =
+  m (NewEpochState era)
+conwayRegisterInitialFundsThenStaking withFileHandle cfg newEpochState = do
   -- We must first register the initial funds, because the stake
   -- information depends on it.
-  resetStakeDistribution
-    . registerDRepsThenDelegs cfg
-    . conwayRegisterInitialAccounts (cfg ^. tcInitialStakingL)
-    . registerInitialStakePools (cfg ^. tcInitialStakingL)
-    . registerInitialFunds cfg
+  newEpochState' <- registerInitialFunds withFileHandle cfg newEpochState
+  pure $
+    resetStakeDistribution $
+      registerDRepsThenDelegs cfg $
+        conwayRegisterInitialAccounts (cfg ^. tcInitialStakingL) $
+          registerInitialStakePools (cfg ^. tcInitialStakingL) newEpochState'
 
--- | Register all staking credentials and apply delegations. Make sure StakePools that are bing
+-- | Register all staking credentials and apply delegations. Make sure StakePools that are being
 -- delegated to are already registered, which can be done with `registerInitialStakePools`.
 conwayRegisterInitialAccounts ::
   forall era.