+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Arbitrary.hs
@@ -607,6 +607,7 @@ instance
     sgGenDelegs <- arbitrary
     sgInitialFunds <- arbitrary
     sgStaking <- arbitrary
+    let sgExtraConfig = Nothing
     pure ShelleyGenesis {..}
 
 instance Arbitrary ShelleyGenesisStaking where